admin_service/requirements.txt
asgiref==3.5.2
attrs==22.1.0
dataclasses-json==0.5.7
Django==4.1.2
django-debug-toolbar==3.7.0
django-split-settings==1.2.0
flake8==4.0.0
iniconfig==1.1.1
marshmallow==3.18.0
marshmallow-enum==1.5.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.4
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
pytz==2022.5
sqlparse==0.4.3
tomli==2.0.1
typing-inspect==0.8.0
typing_extensions==4.4.0
uWSGI==2.0.20
